Rename sqlite-extended to fix overlays
With the original name `sqlite` I hit the same issue I saw on my private
flake: there are packages that depend on the `final` version of `sqlite`
and there appears to be a circular dependency that doesn’t resolve.

Also the outputs of `sqlite-extended` do not match those of `sqlite. In
practice, I could change the outputs of the former to match the latter.
For now, changing the name is fine.