Fix composability of python package overriding
As discussed at NixOS/nixpkgs#44426, `packageOverrides` is not
composable. Haskell packaging has (had?) a similar problem.

The solution, which works fairly elegantly, is to instead use
`python3.pkgs.overrideScope` to define a new set of `pkgs`, and replace
the original `python3.pkgs` with that.

Importantly, the arguments to `packageOverrides` do not need to change,
to be used with `overrideScope`.

Note, also, that `python3.pkgs.overrideScope` corresponds to
`lib.overrideScope'` (with the prime).